Create Combined Package Documentation as Text

This function produces a single text output for a package by processing its documentation (Rd files from package source, or the documentation from alteady installed packages) and vignettes (from *.Rmd, *.qmd, *.md, and*.Rnw` files). The user can provide the function with an installed package name, a package source directory, a tar.gz archive, or a package name to download from CRAN.

Usage

rdd_to_txt(
  pkg,
  file = NULL,
  keep_files = "none",
  cache_path = getOption("rdocdump.cache_path"),
  force_fetch = FALSE
)

Arguments

pkg

A character string specifying the package. This can be:

an installed package name,

a full path to a package source directory,

a full path to a package archive file (tar.gz), or

a package name not installed (which will then be downloaded from CRAN).

file

Optional. Save path for the output text file. If set, the function will return back the path to the file instead of the combined text.

keep_files

A character value controlling whether temporary files should be kept. Possible values are:

"none": Delete both the tar.gz archive and the extracted files (default).

"tgz": Keep only the tar.gz archive.

"extracted": Keep only the extracted files.

"both": Keep both the tar.gz archive and the extracted files.

cache_path

A character string specifying the directory where kept temporary files will be stored. By default, it uses the value of option("rdocdump.cache_path") which sets the cache directory to the temporary directory of the current R session.

force_fetch

logical. If TRUE, the package source will be fetched from CRAN as a tar.gz archive even if the package is already installed locally Default is FALSE.

Value

A single string containing the combined package documentation. If file argument is set, returns path to the file.

Examples

# extract documentation for built-in `stats` package
docs <- rdd_to_txt("stats")

# \donttest{
# extract documentation for rJavaEnv by downloading its source from CRAN
local({
  old_repos <- getOption("repos")
  options(repos = c(CRAN = "https://cran.r-project.org"))
  docs <- rdd_to_txt("rJavaEnv", force_fetch = TRUE)
  lines <- unlist(strsplit(docs, "\n"))
  # Print the first 3 lines
  cat(head(lines, 3), sep = "\n")
  # Print the last 3 lines
  cat(tail(lines, 3), sep = "\n")
  options(repos = old_repos)
  #unlink(getOption("rdocdump.cache_path"), recursive = TRUE)
})
#> Fetching package source from CRAN...
#> DESCRIPTION:
#> Package: rJavaEnv
#> Title: 'Java' Environments for R Projects
#> Now you can remove the package and restart the current `R` session so that `R` picks up the system `Java` (or no `Java`, if it is not installed in the system).
#> 
#> 
# }