Create Combined Package Documentation as Text

This function produces a single text output for a package by processing its Rd files. The package can be provided as an installed package name, a package source directory, a tar.gz archive, or a package name to download from CRAN. (In the future, vignettes will be processed as well.)

Usage

rdd_to_txt(
  pkg,
  file = NULL,
  keep_files = "none",
  cache_path = getOption("rdocdump.cache_path"),
  force_fetch = FALSE
)

Arguments

pkg

A character string specifying the package. This can be:

an installed package name,

a full path to a package source directory,

a full path to a package archive file (tar.gz), or

a package name not installed (which will then be downloaded from CRAN).

file

Optional. Full path for the output text file.

keep_files

Character value controlling whether temporary files should be kept. Possible values are:

"none": Delete both the tar.gz archive and the extracted files (default).

"tgz": Keep only the tar.gz archive.

"extracted": Keep only the extracted files.

"both": Keep both the tar.gz archive and the extracted files.

cache_path

Character string specifying the directory where kept temporary files will be stored. By default, it uses the value of option("rdocdump.cache_path").

force_fetch

Logical. If TRUE, the package source will be fetched from CRAN as a tar.gz archive even if the package is already installed locally. Default is FALSE.

Value

A single string containing the combined package documentation.