Create Combined Package Documentation as Text

This function produces a single text output for a package by processing its documentation (Rd files from the package source, or the documentation from already installed packages) and/or vignettes (from *.Rmd, *.qmd, *.md, and *.Rnw files). The user can provide the function with an installed package name, a package source directory, a tar.gz archive, or a package name to download from CRAN.

Usage

rdd_to_txt(
  pkg,
  file = NULL,
  content = "both",
  force_fetch = FALSE,
  keep_files = "none",
  cache_path = getOption("rdocdump.cache_path")
)

Arguments

pkg

A character string specifying the package. This can be:

an installed package name,

a full path to a package source directory,

a full path to a package archive file (tar.gz), or

a package name not installed (which will then be downloaded from CRAN).

file

Optional. Save path for the output text file. If set, the function will return the path to the file instead of the combined text. Defaults to NULL.

content

A character string specifying which components to include in the output. Possible values are:

"both": Include both Rd documentation and vignettes (default).

"docs": Include only the Rd documentation.

"vignettes": Include only the vignettes.

force_fetch

logical. If TRUE, the package source will be fetched from CRAN as a tar.gz archive even if the package is already installed locally. Default is FALSE.

keep_files

A character value controlling whether temporary files should be kept. Possible values are:

"none": Delete both the tar.gz archive and the extracted files (default).

"tgz": Keep only the tar.gz archive.

"extracted": Keep only the extracted files.

"both": Keep both the tar.gz archive and the extracted files.

cache_path

A character string specifying the directory where kept temporary files will be stored. By default, it uses the value of getOption("rdocdump.cache_path") which sets the cache directory to the temporary directory of the current R session.

Value

A single string containing the combined package documentation. If the file argument is set, returns the path to the file.

Examples

# Extract documentation for built-in `stats` package (both docs and vignettes).
docs <- rdd_to_txt("stats")

# \donttest{
# Extract only documentation for rJavaEnv by downloading its source from CRAN
local({
  old_repos <- getOption("repos")
  options(repos = c(CRAN = "https://cran.r-project.org"))
  docs <- rdd_to_txt("rJavaEnv", force_fetch = TRUE, content = "docs")
  lines <- unlist(strsplit(docs, "\n"))
  # Print the first 3 lines
  cat(head(lines, 3), sep = "\n")
  # Print the last 3 lines
  cat(tail(lines, 3), sep = "\n")
  options(repos = old_repos)
})
#> Fetching package source from CRAN...
#> DESCRIPTION:
#> Package: rJavaEnv
#> Title: 'Java' Environments for R Projects
#>      "17" == java_check_version_rjava(quiet = TRUE)
#>      ## End(Not run)
#>      
# }